pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "phredator"
dynamic = ["version"]
description = "Intelligent NGS Quality Control with automated fix suggestions"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["bioinformatics", "ngs", "qc", "quality-control", "fastqc", "multiqc", "sequencing", "genomics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.20.0",
    "PyYAML>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/phredator"
Repository = "https://github.com/yourusername/phredator"
Issues = "https://github.com/yourusername/phredator/issues"

[project.scripts]
phredator = "phredator.cli.cli:main"

[tool.setuptools]
packages = ["phredator", "phredator.analyzer", "phredator.cli", "phredator.config", "phredator.fixer", "phredator.parser", "phredator.reporter", "phredator.rules", "phredator.utils"]

[tool.setuptools.package-data]
phredator = ["config/*.yaml"]
